Modify each service to use a different logging pattern
Each package that logs has its own interface for the logger to
implement. The interface is focused more on the actions that occur
rather than log messages and the log messages and control of the logger
is controlled in the `diagnostics` package.

This changes each one of the packages except for the tsdb packages to
this new pattern. The tsdb packages are more complicated to convert so
that will be a separate commit that can be reviewed on its own.

This also removes the logger from the execution context in the query
package altogether. It wasn't being used.